test/fixtures/managed_instance_group/requirements.txt
terraform_external_data
apache-libcloud
pycrypto
